Candlepin Spring - Server Fixes (Reasteasy fix, JSProvider fix) [Needs Refactoring]
