+++ server/src/main/java/org/candlepin/ApplicationConfiguration.java
@@ -39,6 +39,8 @@ import org.candlepin.messaging.impl.artemis.ArtemisUtil;
 import org.candlepin.messaging.impl.noop.NoopContextListener;
 import org.candlepin.messaging.impl.noop.NoopSessionFactory;
 import org.candlepin.model.RulesCurator;
+import org.candlepin.policy.js.JsRunnerFactory;
+import org.candlepin.policy.js.JsRunnerRequestCacheFactory;
 import org.candlepin.resteasy.filter.CandlepinSuspendModeFilter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -252,7 +254,7 @@ public class ApplicationConfiguration  implements WebMvcConfigurer  {
 
     @Bean
     public JsRunnerFactory jsRunnerFactory(RulesCurator rulesCurator,
-                                           JsRunnerRequestCacheFactory cacheProvider) {
+        JsRunnerRequestCacheFactory cacheProvider) {
         JsRunnerFactory jsRunnerFactory = new JsRunnerFactory(rulesCurator, cacheProvider);
         return jsRunnerFactory;
     }
