+++ server/src/main/java/org/candlepin/async/tasks/OrphanCleanupJob.java
@@ -14,6 +14,10 @@
  */
 package org.candlepin.async.tasks;
 
+import java.util.List;
+import java.util.Set;
+
+import org.apache.commons.lang3.tuple.Pair;
 import org.candlepin.async.AsyncJob;
 import org.candlepin.async.JobExecutionContext;
 import org.candlepin.async.JobExecutionException;
