+++ server/src/main/java/org/candlepin/resource/RulesResource.java
@@ -28,6 +28,7 @@ import io.swagger.annotations.ApiResponses;
 import io.swagger.annotations.Authorization;
 
 import org.apache.commons.codec.binary.Base64;
+import org.candlepin.policy.js.JsRunnerFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -57,7 +58,7 @@ public class RulesResource {
     private RulesCurator rulesCurator;
     private I18n i18n;
     private EventSink sink;
-    private JsRunnerProvider jsProvider;
+    //private JsRunnerProvider jsProvider;
     /* TODO chnage the instance name to jsFactory */
     private JsRunnerFactory jsProvider;
 
@@ -68,7 +69,7 @@ public class RulesResource {
     //@Inject
     @Autowired
     public RulesResource(RulesCurator rulesCurator,
-        I18n i18n, EventSink sink, JsRunnerProvider jsProvider) {
+        I18n i18n, EventSink sink, JsRunnerFactory jsProvider) {
         this.rulesCurator = rulesCurator;
         this.i18n = i18n;
         this.sink = sink;
