+++ server/src/main/resources/application.properties
@@ -1,0 +1,93 @@
+#
+#  Copyright (c) 2009 - ${YEAR} Red Hat, Inc.
+#
+#  This software is licensed to you under the GNU General Public License,
+#  version 2 (GPLv2). There is NO WARRANTY for this software, express or
+#  implied, including the implied warranties of MERCHANTABILITY or FITNESS
+#  FOR A PARTICULAR PURPOSE. You should have received a copy of GPLv2
+#  along with this software; if not, see
+#  http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
+#
+#  Red Hat trademarks are not licensed under GPLv2. No permission is
+#  granted to use or replicate Red Hat trademarks that are incorporated
+#  in this software or its documentation.
+#
+logging.level.org.springframework.web=debug
+logging.level.org.springframework.boot.autoconfigure=debug
+logging.level.org.springframework=DEBUG
+logging.level.liquibase=DEBUG
+
+spring.guice.dedup=true
+spring.mvc.locale=en
+spring.jackson.serialization.indent_output=true
+spring.autoconfigure.exclude= org.springframework.boot.autoconfigure.jms.activemq.ActiveMQAutoConfiguration, org.springframework.boot.autoconfigure.jms.artemis.ArtemisAutoConfiguration
+server.port=8443
+server.ssl.enabled=true
+server.ssl.key-store-type=PKCS12
+#server.ssl.key-store=/etc/candlepin/certs/myssl.p12
+server.ssl.key-store=/etc/candlepin/certs/keystore
+server.ssl.key-store-password=password
+server.ssl.key-alias=tomcat
+server.ssl.trust-store=/etc/candlepin/certs/keystore
+server.ssl.trust-store-password=password
+security.require-ssl=true
+server.ssl.enabled-protocols=TLSv1.2
+server.ssl.client-auth=want
+#server.ssl.ciphers="SSL_RSA_WITH_3DES_EDE_CBC_SHA,TLS_RSA_WITH_AES_256_CBC_SHA,TLS_ECDH_ECDSA_WITH_AES_128_CBC_SHA,TLS_ECDH_ECDSA_WITH_AES_256_CBC_SHA,TLS_ECDH_RSA_WITH_3DES_EDE_CBC_SHA,TLS_ECDH_RSA_WITH_AES_128_CBC_SHA,TLS_ECDH_RSA_WITH_AES_256_CBC_SHA,TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA"
+
+
+# Database Properties
+spring.datasource.url=${jpa.config.hibernate.connection.url}
+spring.datasource.username=${jpa.config.hibernate.connection.username}
+spring.datasource.password=${jpa.config.hibernate.connection.password}
+spring.datasource.driver-class-name=${jpa.config.hibernate.connection.driver_class}
+spring.datasource.type = com.mchange.v2.c3p0.ComboPooledDataSource
+
+# Hibernate Properties
+spring.jpa.hibernate.ddl-auto=none
+spring.jpa.properties.hibernate.dialect=${jpa.config.hibernate.dialect}
+spring.jpa.hibernate.c3p0.min_size=5
+spring.jpa.hibernate.c3p0.max_size=20
+spring.jpa.hibernate.c3p0.acquire_increment=5
+spring.jpa.hibernate.c3p0.timeout=300
+spring.jpa.hibernate.current_session_context_class=org.springframework.orm.hibernate5.SpringSessionContext
+spring.jpa.properties.hibernate.current_session_context_class=org.springframework.orm.hibernate5.SpringSessionContext
+spring.jpa.hibernate.naming.implicit-strategy=org.hibernate.boot.model.naming.ImplicitNamingStrategyLegacyJpaImpl
+spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
+spring.jpa.properties.hibernate.cache.region.factory_class=org.hibernate.cache.jcache.JCacheRegionFactory
+spring.jpa.properties.hibernate.cache.use_second_level_cache=true
+spring.jpa.properties.hibernate.cache.use_query_cache=true
+spring.jpa.properties.hibernate.javax.cache.missing_cache_strategy=create
+spring.jpa.properties.javax.persistence.sharedCache.mode=ALL
+spring.jpa.show-sql=true
+hibernate.connection.driver_class=org.postgresql.Driver
+#hibernate.dialect=org.hibernate.dialect.PostgreSQL92Dialect
+## spring.jpa.hibernate.ddl-auto=create
+#hibernate.show_sql=true
+spring.jpa.hibernate.driver_class=org.postgresql.Driver
+#hibernate.c3p0.min_size=5
+#hibernate.c3p0.max_size=20
+#hibernate.c3p0.timeout=300
+#hibernate.cache.region.factory_class=org.hibernate.cache.jcache.JCacheRegionFactory
+#hibernate.cache.use_second_level_cache=true
+#hibernate.cache.use_query_cache=true
+#hibernate.javax.cache.missing_cache_strategy=create
+## test period in seconds
+#hibernate.c3p0.idle_test_period=300
+## max_statements should always be 0
+#hibernate.c3p0.max_statements=0
+#hibernate.current_session_context_class=org.springframework.orm.hibernate5.SpringSessionContext
+##Turn Statistics on and log SQL stmts
+#spring.jpa.properties.hibernate.format_sql=true
+#Extensive logging for testing purpose
+# spring.jpa.properties.hibernate.generate_statistics=true
+#logging.level.org.hibernate.type=trace
+#logging.level.org.hibernate.stat=debug
+
+# Liquibase Configuration
+#spring.liquibase.enabled=false
+#spring.liquibase.change-log=classpath:db/changelog/changelog-update.xml
+#spring.liquibase.change-log=classpath:db/changelog/changelog-create.xml
+candlepin.create_database=false
+
+#server.servlet.context-parameters.resteasy.logger.type=SLF4J
